+++ src/Ciphers/ROT8/rot8Encoding.js
