+++ src/alphabet.js
