+++ src/resultParsing.js
